Sync content from opx-sync-2018-06-22_10.4.1.0 to opx-sync-3.0.0_20180803

+++ nbr-mgr/src/nbr_mgr_proc.cpp
@@ -104,14 +104,24 @@ bool nbr_get_if_data(hal_vrf_id_t vrfid, hal_ifindex_t idx, nbr_mgr_intf_entry_t
     return true;
 }
 
-//Return the admin status of interface - true if up, false if down
+//Return the overall admin & oper. status of interface - true if up, false if down
 static bool nbr_check_intf_up(hal_vrf_id_t vrfid, hal_ifindex_t idx) {
     nbr_mgr_intf_entry_t intf;
     if(nbr_get_if_data(vrfid, idx, intf))
-        return (intf.is_admin_up);
+        return ((intf.is_admin_up) && (intf.is_oper_up));
 
     return false;
 }
+
+//Return the oper status of interface - true if up, false if down
+static bool nbr_check_intf_oper_up(hal_vrf_id_t vrfid, hal_ifindex_t idx) {
+    nbr_mgr_intf_entry_t intf;
+    if(nbr_get_if_data(vrfid, idx, intf))
+        return (intf.is_oper_up);
+
+    return false;
+}
+
 
 //Nbr process Class definitions
 void nbr_process::process_nbr_entries(void)
@@ -536,11 +546,6 @@ bool nbr_process::delete_nbr_instance(nbr_ip_db_type& nbr_db, hal_vrf_id_t vrf, 
     return nbr_db_remove(nbr_db, vrf, key);
 }
 
-inline bool nbr_check_for_duplicate(nbr_mgr_intf_entry_t& intf1, nbr_mgr_intf_entry_t& intf2) {
-    if((intf1.is_admin_up == intf2.is_admin_up) && (intf1.vlan_id == intf2.vlan_id)) return true;
-    return false;
-}
-
 bool nbr_process::nbr_update_intf_info(nbr_mgr_intf_entry_t& intf) {
     NBR_MGR_LOG_INFO("PROC", "Intf update VRF-id:%lu is-del:%d, if_index:%d, flags:%d is_admin_up:%d,"
                      "is_bridge:%d vlan-id:%d", intf.vrfid, intf.is_op_del, intf.if_index,
@@ -551,11 +556,16 @@ bool nbr_process::nbr_update_intf_info(nbr_mgr_intf_entry_t& intf) {
 }
 
 bool nbr_process::nbr_proc_intf_msg(nbr_mgr_intf_entry_t& intf) {
-    NBR_MGR_LOG_INFO("PROC", "Intf VRF-id:%lu is-del:%d, if_index:%d, flags:%d is_admin_up:%d,"
-                     "is_bridge:%d vlan-id:%d", intf.vrfid, intf.is_op_del, intf.if_index,
-                     intf.flags, intf.is_admin_up, intf.is_bridge, intf.vlan_id);
+    NBR_MGR_LOG_INFO("PROC", "Intf VRF-id:%lu(%s) is-del:%d, if_index:%d(%s), msg-type:%s "
+                    "update msg flags:%d is_admin_up:%d is_oper_up:%d,"
+                    "is_bridge:%d vlan-id:%d", intf.vrfid, intf.vrf_name, intf.is_op_del, intf.if_index,
+                    intf.if_name,
+                    ((intf.flags == NBR_MGR_INTF_VLAN_MSG) ? "VLAN" :
+                     ((intf.flags == NBR_MGR_INTF_ADMIN_MSG) ? "Admin" : "Oper")),
+                    intf.flags, intf.is_admin_up, intf.is_oper_up, intf.is_bridge, intf.vlan_id);
 
     if(!intf.is_op_del) {
+        bool intf_update_req = true;
         auto vrf_itr = neighbor_if_db.find(intf.vrfid);
         if (vrf_itr != neighbor_if_db.end()) {
             auto &intf_list = vrf_itr->second;
@@ -563,26 +573,54 @@ bool nbr_process::nbr_proc_intf_msg(nbr_mgr_intf_entry_t& intf) {
             auto itr = intf_list.find(intf.if_index);
             stats.intf_add_msg_cnt++;
             if(itr != intf_list.end()) {
+                intf_update_req = false;
                 auto& intf_db = itr->second;
-                /* Update the VLAN-id and return. */
-                if (intf.flags == NBR_MGR_INTF_VLAN_MSG) {
-                    intf_db.vlan_id = intf.vlan_id;
-                    return true;
-                    /* Admin status changed and VLAN already present,
-                     * copy the VLAN-id and handle Admin status change */
-                } else if ((intf.flags == NBR_MGR_INTF_ADMIN_MSG) &&
-                           (intf_db.vlan_id)) {
-                    intf.vlan_id = intf_db.vlan_id;
+
+                NBR_MGR_LOG_INFO("PROC-INTF-DB", "Current info - Intf VRF-id:%lu if_index:%d, "
+                                 "is_admin_up:%d is_oper_up:%d, is_bridge:%d vlan-id:%d",
+                                 intf_db.vrfid, intf_db.if_index, intf_db.is_admin_up, intf_db.is_oper_up,
+                                 intf_db.is_bridge, intf_db.vlan_id);
+
+                if (intf.flags & NBR_MGR_INTF_VLAN_MSG) {
+                    if (intf_db.vlan_id == intf.vlan_id) {
+                        intf.flags &= ~NBR_MGR_INTF_VLAN_MSG;
+                    } else {
+                        intf_db.vlan_id = intf.vlan_id;
+                    }
                 }
-                /* If both admin and VLAN are changed, just handle it here. */
-                if(nbr_check_for_duplicate(itr->second, intf))
-                    return false;
-                intf.parent_or_child_vrfid = intf_db.parent_or_child_vrfid;
-                intf.parent_or_child_if_index = intf_db.parent_or_child_if_index;
+                if (intf.flags & NBR_MGR_INTF_OPER_MSG) {
+                    if (intf_db.is_oper_up == intf.is_oper_up) {
+                        intf.flags &= ~NBR_MGR_INTF_OPER_MSG;
+                    } else {
+                        intf_db.is_oper_up = intf.is_oper_up;
+                    }
+                }
+                if (intf.flags & NBR_MGR_INTF_ADMIN_MSG) {
+                    if (intf_db.is_admin_up == intf.is_admin_up) {
+                        intf.flags &= ~NBR_MGR_INTF_ADMIN_MSG;
+                    } else {
+                        intf_db.is_admin_up = intf.is_admin_up;
+                        nbr_mgr_notify_intf_status (NBR_MGR_OP_CREATE, itr->second);
+                    }
+                }
+                if (intf.flags == 0) {
+                    NBR_MGR_LOG_INFO("PROC-INTF-DB", "Duplicate intf. info update - Intf VRF-id:%lu if_index:%d, "
+                                     "is_admin_up:%d is_oper_up:%d, is_bridge:%d vlan-id:%d",
+                                     intf_db.vrfid, intf_db.if_index, intf_db.is_admin_up, intf_db.is_oper_up,
+                                     intf_db.is_bridge, intf_db.vlan_id);
+                    return true;
+                }
+
+                NBR_MGR_LOG_INFO("PROC-INTF-DB", "Updated info - Intf VRF-id:%lu if_index:%d, "
+                                 "is_admin_up:%d is_oper_up:%d, is_bridge:%d vlan-id:%d",
+                                 intf_db.vrfid, intf_db.if_index, intf_db.is_admin_up, intf_db.is_oper_up,
+                                 intf_db.is_bridge, intf_db.vlan_id);
             }
         }
-        nbr_mgr_notify_intf_status (NBR_MGR_OP_CREATE, intf);
-        neighbor_if_db[intf.vrfid][intf.if_index] = intf;
+        if (intf_update_req) {
+            nbr_mgr_notify_intf_status (NBR_MGR_OP_CREATE, intf);
+            neighbor_if_db[intf.vrfid][intf.if_index] = intf;
+        }
     } else {
         nbr_mgr_notify_intf_status (NBR_MGR_OP_DELETE, intf);
         auto vrf_itr = neighbor_if_db.find(intf.vrfid);
@@ -612,12 +650,29 @@ bool nbr_process::nbr_proc_intf_msg(nbr_mgr_intf_entry_t& intf) {
         stats.intf_del_msg_cnt++;
     }
 
+    /* If interface name is not empty and oper status is down, flush the neighbors */
+    if ((intf.if_name[0] != '\0') && (intf.flags & NBR_MGR_INTF_OPER_MSG)
+        && (intf.is_oper_up == false)) {
+        auto v_itr = neighbor_if_nbr_db.find(intf.vrfid);
+        if (v_itr != neighbor_if_nbr_db.end()) {
+            /* VRF->Intf-list */
+            auto &if_list = v_itr->second;
+
+            /* Walk interface-> Nbrs in the VRF */
+            const auto itr = if_list.find(intf.if_index);
+            if(itr != if_list.end()) {
+                nbr_mgr_os_neigh_flush(intf.vrf_name, HAL_INET4_FAMILY, intf.if_name);
+                nbr_mgr_os_neigh_flush(intf.vrf_name, HAL_INET6_FAMILY, intf.if_name);
+            }
+        }
+    }
+
     nbr_if_db_walk(intf.vrfid, intf.if_index, [&](nbr_data const * n_ptr) mutable {
         const_cast<nbr_data *>(n_ptr)->handle_if_state_change(intf);
         nbr_ip_db_type& nbr_db = neighbor_db(n_ptr->get_family());
 
         //If interface is deleted or down, delete the nbr instances
-        if(intf.is_op_del || !intf.is_admin_up) {
+        if(intf.is_op_del || ((intf.flags & NBR_MGR_INTF_ADMIN_MSG) && !intf.is_admin_up)) {
             std::string key = nbr_get_key(n_ptr->get_if_index(), n_ptr->get_ip_addr());
             this->delete_nbr_instance(nbr_db, n_ptr->get_vrf_id(), key);
         }
@@ -974,6 +1029,7 @@ bool nbr_data::process_nbr_data(nbr_mgr_nbr_entry_t& entry) {
                      * refresh again */
                     if (m_refresh_cnt) {
                         reset_refresh_cnt();
+                        nbr_stats.refresh_for_more_flushes++;
                         trigger_refresh();
                     } else {
                         m_prev_refresh_for_mac_learn_retry_cnt = m_refresh_for_mac_learn_retry_cnt;
@@ -1064,7 +1120,7 @@ bool nbr_data::process_nbr_data(nbr_mgr_nbr_entry_t& entry) {
                     m_retry_cnt++;
                     /* If the MAC is not learnt, it could be because of
                      * the MAC learning disable on the L3 interface, wait for APP to program the MAC */
-                    rc = trigger_delay_refresh();
+                    trigger_delay_refresh();
                     NBR_MGR_LOG_INFO("PROC", "MAC data is not present, vrf-id:%d(%s), family:%d,"
                                      "Nbr:%s, if_index:%d (ll-intf:%d), phy_idx:%d, status:%d,"
                                      "flags:0x%x entry type:%d status:%lu flags:%lu",
@@ -1086,7 +1142,21 @@ bool nbr_data::process_nbr_data(nbr_mgr_nbr_entry_t& entry) {
             m_flags &= ~NBR_MGR_NBR_REFRESH_FOR_MAC_LEARN;
             m_prev_refresh_for_mac_learn_retry_cnt = m_refresh_for_mac_learn_retry_cnt;
             m_refresh_for_mac_learn_retry_cnt = 0;
-            if (m_flags & NBR_MGR_NBR_REFRESH) {
+            /* If the interface is operationally down, dont trigger resolution for regular hosts,
+             * after oper.up, with the traffic, hosts are expected to resolve in kernel. */
+            if (nbr_check_intf_oper_up(m_vrf_id, m_ifindex) == false) {
+                m_flags &= ~NBR_MGR_NBR_REFRESH;
+                reset_refresh_cnt();
+                NBR_MGR_LOG_INFO("PROC", "Pub vrf-id:%d(%s), family:%d,"
+                                "Nbr:%s, if_index:%d (ll-intf:%d), phy_idx:%d, status:%d,"
+                                "flags:0x%x entry type:%d status:%lu flags:%lu",
+                                m_vrf_id, m_vrf_name.c_str(), m_family, ip.c_str(), m_ifindex,
+                                m_parent_if, entry.mbr_if_index, m_status,
+                                m_flags, entry.msg_type, entry.status, entry.flags);
+
+                nbr_stats.failed_handle_skip_oper_down++;
+                rc = publish_entry(NBR_MGR_OP_CREATE, entry);
+            } else if (m_flags & NBR_MGR_NBR_REFRESH) {
                 /* Since there can be a topology settling issue,
                  * try the refresh/resolve for retry_cnt and
                  * if fails, remove the nbr entry in the NPU to lift
@@ -1127,7 +1197,7 @@ bool nbr_data::process_nbr_data(nbr_mgr_nbr_entry_t& entry) {
                 m_flags |= NBR_MGR_NBR_REFRESH;
                 trigger_resolve();
             } else if (m_flags & NBR_MGR_NBR_RESOLVE) {
-                trigger_resolve();
+                rc = trigger_resolve();
                 rc = publish_entry(NBR_MGR_OP_CREATE, entry);
             } else {
                 /* If Nbr flushed, delete the nbr entry from the NPU.*/
@@ -1214,8 +1284,8 @@ void nbr_data::populate_nbr_entry(nbr_mgr_nbr_entry_t& entry) const {
 
 bool nbr_data::handle_fdb_change(nbr_mgr_evt_type_t evt, unsigned long status) const {
     std::string ip = nbr_ip_addr_string (m_ip_addr);
-    NBR_MGR_LOG_DEBUG("PROC-NL-MSG", "Handle FDB change for %s if-index:%d",
-                      ip.c_str(), m_ifindex);
+    NBR_MGR_LOG_DEBUG("PROC-NL-MSG", "Handle FDB change for %s if-index:%d evt:%d status:%d",
+                      ip.c_str(), m_ifindex, evt, status);
 
     /* On FDB del (may be because of STP topology change), refresh the ARPs/Nbrs
      * to re-learn on correct port using ARP/NA reply incase of uni-directional
@@ -1228,6 +1298,11 @@ bool nbr_data::handle_fdb_change(nbr_mgr_evt_type_t evt, unsigned long status) c
         if (!(m_status & NBR_MGR_NUD_PERMANENT) &&
             !(m_status & NBR_MGR_NUD_FAILED) &&
             !(m_status & NBR_MGR_NUD_INCOMPLETE)) {
+            if ((evt == NBR_MGR_NBR_DEL) &&
+                (nbr_check_intf_oper_up(m_vrf_id, m_ifindex) == false)) {
+                /* Oper. down case, dont refresh the ARP/Nbr. */
+                return true;
+            }
             m_flags |= NBR_MGR_NBR_REFRESH;
             nbr_stats.mac_trig_refresh++;
             return trigger_refresh();
@@ -1284,16 +1359,28 @@ bool nbr_data::handle_fdb_change(nbr_mgr_evt_type_t evt, unsigned long status) c
 bool nbr_data::handle_if_state_change(nbr_mgr_intf_entry_t& intf) {
 
     std::string ip = nbr_ip_addr_string (m_ip_addr);
-    NBR_MGR_LOG_DEBUG("PROC-IF-CHG", "Handle IF state change for %s if-index:%d "
-                      "admin %d, m_flags 0x%x, m_status 0x%x ",
-                      ip.c_str(), intf.if_index, intf.is_admin_up, m_flags, m_status);
-    if(intf.is_admin_up) {
+    NBR_MGR_LOG_INFO("PROC-IF-CHG", "Handle IF state change for %s if-index:%d "
+                     "admin %d, oper %d m_flags 0x%x, m_status 0x%x ",
+                     ip.c_str(), intf.if_index, intf.is_admin_up, intf.is_oper_up, m_flags, m_status);
+    bool trig_nbr_resolve = false;
+    if (intf.flags & NBR_MGR_INTF_ADMIN_MSG) {
+        if (intf.is_admin_up) {
+            trig_nbr_resolve = true;
+        } else {
+            m_status = NBR_MGR_NUD_NONE;
+        }
+    }
+    if (intf.flags & NBR_MGR_INTF_OPER_MSG) {
+        if (intf.is_oper_up) {
+            trig_nbr_resolve = true;
+        }
+    }
+    if (trig_nbr_resolve) {
         if((m_flags & NBR_MGR_NBR_RESOLVE) &&
-            !(m_status & NBR_MGR_NUD_REACHABLE) &&
-            !(m_status & NBR_MGR_NUD_PERMANENT))
+           !(m_status & NBR_MGR_NUD_REACHABLE) &&
+           !(m_status & NBR_MGR_NUD_PERMANENT))
             trigger_resolve();
-    } else
-        m_status = NBR_MGR_NUD_NONE;
+    }
 
     return true;
 }
@@ -1342,6 +1429,24 @@ bool nbr_data::trigger_resolve() const {
     nbr_stats.resolve_cnt++;
     /* Resolve the neighbor */
     return nbr_mgr_nbr_resolve(NBR_MGR_NL_RESOLVE_MSG, &nbr);
+}
+
+bool nbr_data::trigger_delay_resolve() const {
+    nbr_mgr_nbr_entry_t nbr;
+    std::string ip = nbr_ip_addr_string (m_ip_addr);
+    NBR_MGR_LOG_DEBUG("PROC-NL-MSG", "Resolve the neighbor %s if-index:%d",
+                     ip.c_str(), m_ifindex);
+
+    memset(&nbr, 0, sizeof(nbr));
+    populate_nbr_entry(nbr);
+
+    //Skip resolve if the interface is admin down/not exist
+    if(!nbr_check_intf_up(nbr.vrfid, nbr.if_index))
+        return false;
+
+    nbr_stats.delay_resolve_cnt++;
+    /* Resolve the neighbor */
+    return nbr_mgr_nbr_resolve(NBR_MGR_NL_DELAY_RESOLVE_MSG, &nbr);
 }
 
 bool nbr_data::trigger_refresh() const {
@@ -1404,36 +1509,6 @@ bool nbr_data::trigger_delay_refresh() const {
     return nbr_mgr_nbr_resolve(NBR_MGR_NL_DELAY_REFRESH_MSG, &nbr);
 }
 
-bool nbr_data::trigger_delay_resolve() const {
-    nbr_mgr_nbr_entry_t nbr;
-
-    /* Dont refresh for static ARP entry */
-    if (m_status & NBR_MGR_NUD_PERMANENT)
-        return false;
-
-    std::string ip = nbr_ip_addr_string (m_ip_addr);
-    if (m_mac_data_ptr == NULL) {
-        NBR_MGR_LOG_INFO("PROC", "MAC is NULL, refresh failed for Nbr %s",
-                        ip.c_str());
-        return false;
-    }
-    NBR_MGR_LOG_DEBUG("PROC-NL-MSG", "Delay resolve the neighbor %s MAC:%s if-index:%d",
-                     ip.c_str(), m_mac_data_ptr->get_mac_addr().c_str(), m_ifindex);
-    memset(&nbr, 0, sizeof(nbr));
-    populate_nbr_entry(nbr);
-
-    //Skip refresh if the interface is admin down/not exist
-    if(!nbr_check_intf_up(nbr.vrfid, nbr.if_index))
-        return false;
-
-    if (!(m_flags & NBR_MGR_MAC_NOT_PRESENT))
-        m_flags |= NBR_MGR_NBR_REFRESH;
-
-    nbr_stats.delay_resolve_cnt++;
-    /* Refresh the neighbor */
-    return nbr_mgr_nbr_resolve(NBR_MGR_NL_DELAY_RESOLVE_MSG, &nbr);
-}
-
 bool nbr_data::trigger_refresh_for_mac_learn() const {
     nbr_mgr_nbr_entry_t nbr;
 
@@ -1461,6 +1536,32 @@ bool nbr_data::trigger_refresh_for_mac_learn() const {
     nbr_stats.hw_mac_learn_refresh_cnt++;
     /* Refresh the neighbor */
     return nbr_mgr_nbr_resolve(NBR_MGR_NL_DELAY_REFRESH_MSG, &nbr);
+}
+
+bool nbr_data::trigger_instant_refresh() const {
+    nbr_mgr_nbr_entry_t nbr;
+
+    /* Dont refresh for static ARP entry */
+    if (m_status & NBR_MGR_NUD_PERMANENT)
+        return false;
+
+    std::string ip = nbr_ip_addr_string (m_ip_addr);
+    if (m_mac_data_ptr == NULL) {
+        NBR_MGR_LOG_INFO("PROC", "MAC is NULL, refresh failed for Nbr %s",
+                        ip.c_str());
+        return false;
+    }
+    NBR_MGR_LOG_INFO("PROC-NL-MSG", "Instant refresh the neighbor %s MAC:%s if-index:%d",
+                     ip.c_str(), m_mac_data_ptr->get_mac_addr().c_str(), m_ifindex);
+    memset(&nbr, 0, sizeof(nbr));
+    populate_nbr_entry(nbr);
+
+    if (!(m_flags & NBR_MGR_MAC_NOT_PRESENT))
+        m_flags |= NBR_MGR_NBR_REFRESH;
+
+    nbr_stats.instant_refresh_cnt++;
+    /* Refresh the neighbor */
+    return nbr_mgr_nbr_resolve(NBR_MGR_NL_INSTANT_REFRESH_MSG, &nbr);
 }
